IPIMS_Patient_Care_System-develop/requirements.txt
Django==1.8.5
django-crispy-forms==1.4.0
django-phonenumber-field==0.7.2
phonenumbers==7.1.0